[TSS-52] Pull out asserts
 - refactor: pull out extra assertions from the NSubstitute bridge class
   Fluent into new classes
 - refactor: rename AssertionMatcher -> FluentAssertionMatcher to
   clearly state what it is - a (argument) matcher for
   FluentAssertions-style asserts
