+++ FluentAssertionsEx/FluentAssertionsEx.csproj
@@ -52,12 +52,16 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Assertions\CancellationTokenAssertions.cs" />
+    <Compile Include="Assertions\TaskAssertions.cs" />
+    <Compile Include="Extensions\FluentAssertionsExtensions.cs" />
     <Compile Include="Fluent.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
